+++ src/pyatmo/modules/__init__.py
@@ -1,6 +1,21 @@
 """Expose submodules."""
 from .base_class import Place
-from .bticino import BNCX, BNDL, BNEU, BNSL
+from .bticino import (
+    BNAB,
+    BNAS,
+    BNCS,
+    BNCX,
+    BNDL,
+    BNEU,
+    BNFC,
+    BNIL,
+    BNMH,
+    BNMS,
+    BNSL,
+    BNTH,
+    BNTR,
+    BNXM,
+)
 from .idiamant import NBG, NBO, NBR, NBS
 from .legrand import (
     EBU,
@@ -15,6 +30,7 @@ from .legrand import (
     NLFN,
     NLG,
     NLIS,
+    NLJ,
     NLL,
     NLLF,
     NLLM,
@@ -23,6 +39,7 @@ from .legrand import (
     NLP,
     NLPBS,
     NLPC,
+    NLPD,
     NLPM,
     NLPO,
     NLPS,
@@ -64,11 +81,21 @@ from .smarther import BNS
 from .somfy import TPSRS
 
 __all__ = [
+    "BNMS",
+    "BNAS",
+    "BNAB",
+    "BNMH",
+    "BNTH",
+    "BNFC",
+    "BNTR",
+    "BNXM",
+    "BNCS",
     "BNCX",
     "BNDL",
     "BNEU",
     "BNS",
     "BNSL",
+    "BNIL",
     "Camera",
     "Dimmer",
     "Location",
@@ -106,6 +133,7 @@ __all__ = [
     "NLP",
     "NLPBS",
     "NLPC",
+    "NLPD",
     "NLPM",
     "NLPO",
     "NLPS",
@@ -133,4 +161,5 @@ __all__ = [
     "TPSRS",
     "NLAS",
     "NLTS",
+    "NLJ",
 ]