+++ src/pyatmo/modules/bticino.py
@@ -22,3 +22,43 @@ class BNCX(Module):
 
 class BNEU(Module):
     """BTicino external unit."""
+
+
+class BNCS(Module):
+    """BTicino camera."""
+
+
+class BNXM(Module):
+    """BTicino X meter."""
+
+
+class BNMS(Module):
+    """BTicino motorized shade."""
+
+
+class BNAS(Module):
+    """BTicino automatic shutter."""
+
+
+class BNAB(Module):
+    """BTicino automatic blind."""
+
+
+class BNMH(Module):
+    """BTicino automatic blind."""
+
+
+class BNTH(Module):
+    """BTicino thermostat."""
+
+
+class BNFC(Module):
+    """BTicino fan coil."""
+
+
+class BNTR(Module):
+    """BTicino radiator thermostat."""
+
+
+class BNIL(Switch):
+    """BTicino itelligent light."""