+++ src/pyatmo/modules/legrand.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 import logging
 
 from pyatmo.modules.module import (
+    BatteryMixin,
     ContactorMixin,
     Dimmer,
     EnergyMixin,
@@ -24,15 +25,15 @@ LOG = logging.getLogger(__name__)
 # pylint: disable=R0901
 
 
-class NLG(FirmwareMixin, Module):
+class NLG(FirmwareMixin, OffloadMixin, WifiMixin, Module):
     """Legrand gateway."""
 
 
-class NLT(FirmwareMixin, Module):
+class NLT(FirmwareMixin, BatteryMixin, Module):
     """Legrand global remote control."""
 
 
-class NLP(Switch):
+class NLP(Switch, HistoryMixin, PowerMixin, OffloadMixin, Module):
     """Legrand plug."""
 
 
@@ -73,7 +74,7 @@ class NLIS(Switch):
 
 
 class NLD(Dimmer):
-    """Legrand Double On/Off dimmer remote"""
+    """Legrand Double On/Off dimmer remote."""
 
 
 class NLL(FirmwareMixin, EnergyMixin, WifiMixin, SwitchMixin, Module):
@@ -104,7 +105,7 @@ class NLPS(FirmwareMixin, PowerMixin, EnergyMixin, Module):
     """Legrand / BTicino smart load shedder."""
 
 
-class NLC(FirmwareMixin, SwitchMixin, Module):
+class NLC(FirmwareMixin, SwitchMixin, HistoryMixin, PowerMixin, OffloadMixin, Module):
     """Legrand / BTicino cable outlet."""
 
 
@@ -154,3 +155,11 @@ class EBU(Module):
 
 class NLTS(Module):
     """NLTS motion sensor."""
+
+
+class NLPD(FirmwareMixin, SwitchMixin, Module):
+    """NLPD dry contact."""
+
+
+class NLJ(FirmwareMixin, RfMixin, ShutterMixin, Module):
+    """Legrand garage door opener."""