+++ src/pyatmo/thermostat.py
@@ -1,9 +1,9 @@
 """Support for Netatmo energy devices (relays, thermostats and valves)."""
 from __future__ import annotations
 
-import logging
 from abc import ABC
 from collections import defaultdict
+import logging
 from typing import Any
 from warnings import warn
 
@@ -36,6 +36,7 @@ class AbstractHomeData(ABC):
 
     def process(self) -> None:
         """Process data from API."""
+
         self.homes = {d["id"]: d for d in self.raw_data}
 
         for item in self.raw_data:
@@ -70,25 +71,29 @@ class AbstractHomeData(ABC):
 
     def _get_selected_schedule(self, home_id: str) -> dict:
         """Get the selected schedule for a given home ID."""
+
         return next(
             (
                 value
                 for value in self.schedules.get(home_id, {}).values()
-                if "selected" in value.keys()
+                if "selected" in value
             ),
             {},
         )
 
     def get_hg_temp(self, home_id: str) -> float | None:
         """Return frost guard temperature value."""
+
         return self._get_selected_schedule(home_id).get("hg_temp")
 
     def get_away_temp(self, home_id: str) -> float | None:
         """Return the configured away temperature value."""
+
         return self._get_selected_schedule(home_id).get("away_temp")
 
     def get_thermostat_type(self, home_id: str, room_id: str) -> str | None:
         """Return the thermostat type of the room."""
+
         return next(
             (
                 module.get("type")
@@ -100,6 +105,7 @@ class AbstractHomeData(ABC):
 
     def is_valid_schedule(self, home_id: str, schedule_id: str):
         """Check if valid schedule."""
+
         schedules = (
             self.schedules[home_id][s]["id"] for s in self.schedules.get(home_id, {})
         )
@@ -110,15 +116,13 @@ class HomeData(AbstractHomeData):
     """Class of Netatmo energy devices."""
 
     def __init__(self, auth: NetatmoOAuth2) -> None:
-        """Initialize the Netatmo home data.
+        """Initialize the Netatmo home data."""
 
-        Arguments:
-            auth {NetatmoOAuth2} -- Authentication information with valid access token
-        """
         self.auth = auth
 
     def update(self) -> None:
         """Fetch and process data from API."""
+
         resp = self.auth.post_api_request(endpoint=GETHOMESDATA_ENDPOINT)
 
         self.raw_data = extract_raw_data(resp.json(), "homes")
@@ -126,6 +130,7 @@ class HomeData(AbstractHomeData):
 
     def switch_home_schedule(self, home_id: str, schedule_id: str) -> Any:
         """Switch the schedule for a give home ID."""
+
         if not self.is_valid_schedule(home_id, schedule_id):
             raise NoSchedule(f"{schedule_id} is not a valid schedule id")
 
@@ -141,15 +146,13 @@ class AsyncHomeData(AbstractHomeData):
     """Class of Netatmo energy devices."""
 
     def __init__(self, auth: AbstractAsyncAuth) -> None:
-        """Initialize the Netatmo home data.
+        """Initialize the Netatmo home data."""
 
-        Arguments:
-            auth {AbstractAsyncAuth} -- Authentication information with valid access token
-        """
         self.auth = auth
 
     async def async_update(self):
         """Fetch and process data from API."""
+
         resp = await self.auth.async_post_api_request(endpoint=GETHOMESDATA_ENDPOINT)
 
         assert not isinstance(resp, bytes)
@@ -158,6 +161,7 @@ class AsyncHomeData(AbstractHomeData):
 
     async def async_switch_home_schedule(self, home_id: str, schedule_id: str) -> None:
         """Switch the schedule for a give home ID."""
+
         if not self.is_valid_schedule(home_id, schedule_id):
             raise NoSchedule(f"{schedule_id} is not a valid schedule id")
 
@@ -179,6 +183,7 @@ class AbstractHomeStatus(ABC):
 
     def process(self) -> None:
         """Process data from API."""
+
         for room in self.raw_data.get("rooms", []):
             self.rooms[room["id"]] = room
 
@@ -194,6 +199,7 @@ class AbstractHomeStatus(ABC):
 
     def get_room(self, room_id: str) -> dict:
         """Return room data for a given room id."""
+
         for value in self.rooms.values():
             if value["id"] == room_id:
                 return value
@@ -202,6 +208,7 @@ class AbstractHomeStatus(ABC):
 
     def get_thermostat(self, room_id: str) -> dict:
         """Return thermostat data for a given room id."""
+
         for value in self.thermostats.values():
             if value["id"] == room_id:
                 return value
@@ -210,6 +217,7 @@ class AbstractHomeStatus(ABC):
 
     def get_relay(self, room_id: str) -> dict:
         """Return relay data for a given room id."""
+
         for value in self.relays.values():
             if value["id"] == room_id:
                 return value
@@ -218,6 +226,7 @@ class AbstractHomeStatus(ABC):
 
     def get_valve(self, room_id: str) -> dict:
         """Return valve data for a given room id."""
+
         for value in self.valves.values():
             if value["id"] == room_id:
                 return value
@@ -226,18 +235,22 @@ class AbstractHomeStatus(ABC):
 
     def set_point(self, room_id: str) -> float | None:
         """Return the setpoint of a given room."""
+
         return self.get_room(room_id).get("therm_setpoint_temperature")
 
     def set_point_mode(self, room_id: str) -> str | None:
         """Return the setpointmode of a given room."""
+
         return self.get_room(room_id).get("therm_setpoint_mode")
 
     def measured_temperature(self, room_id: str) -> float | None:
         """Return the measured temperature of a given room."""
+
         return self.get_room(room_id).get("therm_measured_temperature")
 
     def boiler_status(self, module_id: str) -> bool | None:
         """Return the status of the boiler status."""
+
         return self.get_thermostat(module_id).get("boiler_status")
 
 
@@ -245,17 +258,14 @@ class HomeStatus(AbstractHomeStatus):
     """Class of the Netatmo home status."""
 
     def __init__(self, auth: NetatmoOAuth2, home_id: str):
-        """Initialize the Netatmo home status.
+        """Initialize the Netatmo home status."""
 
-        Arguments:
-            auth {NetatmoOAuth2} -- Authentication information with a valid access token
-            home_id {str} -- ID for targeted home
-        """
         self.auth = auth
         self.home_id = home_id
 
     def update(self) -> None:
         """Fetch and process data from API."""
+
         resp = self.auth.post_api_request(
             endpoint=GETHOMESTATUS_ENDPOINT,
             params={"home_id": self.home_id},
@@ -271,6 +281,7 @@ class HomeStatus(AbstractHomeStatus):
         schedule_id: str | None = None,
     ) -> str | None:
         """Set thermotat mode."""
+
         post_params = {"home_id": self.home_id, "mode": mode}
         if end_time is not None and mode in {"hg", "away"}:
             post_params["endtime"] = str(end_time)
@@ -291,6 +302,7 @@ class HomeStatus(AbstractHomeStatus):
         end_time: int | None = None,
     ) -> str | None:
         """Set room themperature set point."""
+
         post_params = {"home_id": self.home_id, "room_id": room_id, "mode": mode}
         # Temp and endtime should only be sent when mode=='manual', but netatmo api can
         # handle that even when mode == 'home' and these settings don't make sense
@@ -310,17 +322,14 @@ class AsyncHomeStatus(AbstractHomeStatus):
     """Class of the Netatmo home status."""
 
     def __init__(self, auth: AbstractAsyncAuth, home_id: str):
-        """Initialize the Netatmo home status.
+        """Initialize the Netatmo home status."""
 
-        Arguments:
-            auth {AbstractAsyncAuth} -- Authentication information with a valid access token
-            home_id {str} -- ID for targeted home
-        """
         self.auth = auth
         self.home_id = home_id
 
     async def async_update(self) -> None:
         """Fetch and process data from API."""
+
         resp = await self.auth.async_post_api_request(
             endpoint=GETHOMESTATUS_ENDPOINT,
             params={"home_id": self.home_id},
@@ -337,6 +346,7 @@ class AsyncHomeStatus(AbstractHomeStatus):
         schedule_id: str | None = None,
     ) -> str | None:
         """Set thermotat mode."""
+
         post_params = {"home_id": self.home_id, "mode": mode}
         if end_time is not None and mode in {"hg", "away"}:
             post_params["endtime"] = str(end_time)
@@ -359,6 +369,7 @@ class AsyncHomeStatus(AbstractHomeStatus):
         end_time: int | None = None,
     ) -> str | None:
         """Set room themperature set point."""
+
         post_params = {"home_id": self.home_id, "room_id": room_id, "mode": mode}
         # Temp and endtime should only be sent when mode=='manual', but netatmo api can
         # handle that even when mode == 'home' and these settings don't make sense